requirements.txt
PyYAML==6.0.1
streamlit==1.32.2
watchdog==4.0.0
